requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==3.2.20
django-apscheduler==0.6.2
djangorestframework==3.14.0
drf-extensions==0.7.1
exceptiongroup==1.1.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
icalendar==5.0.7
idna==3.4
importlib-metadata==6.7.0
lxml==4.9.3
mysqlclient==2.1.1
platformdirs==3.8.1
pycryptodome==3.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-cache==1.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzlocal==5.0.1
url-normalize==1.4.3
urllib3==2.0.3
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
